refactor: organize file structure (still errors to fix)
