+++ sattern/src/main.py
@@ -1,8 +1,8 @@
-import sattern.src.tools.api as api 
+import sattern.src.api as api 
 from sattern.src.metrics.combine import combine
-from sattern.src.tools.display import display
-from sattern.src.tools.llm import run_llm
-from sattern.src.tools.trader import portfolio
+from sattern.src.display import display
+from sattern.src.llm import run_llm
+from sattern.src.trader import portfolio
 from typing import Dict
 import pandas as pd
 from datetime import datetime, timedelta, timezone
