+++ sattern/src/trader.py
